+++ packages/sdk/src/plugins/schema/plugin-sdk/index.ts
@@ -3,4 +3,14 @@ export * from "../plugin";
 export * from "../ipc-envelope";
 export * from "../icons";
 export { PreparedPluginRuntime } from "../plugin/execution";
-export { SETTINGS_API_VERSION, PLUGIN_SCHEMA_VERSION } from "../constants";
+export {
+  SETTINGS_API_VERSION,
+  PLUGIN_SCHEMA_VERSION,
+  PLUGIN_RELEASE_FILES,
+  PLUGIN_PUBLISH_AUDIENCE,
+  PLUGIN_CONFIG_MAX_BYTES,
+  PLUGIN_PUBLISH_MAX_BYTES,
+} from "../constants";
+
+export * from "../plugin-publishing";
+export * from "../plugin-version";